+++ xarray/backends/h5netcdf_.py
@@ -2,7 +2,6 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 import functools
-import warnings
 
 from .. import Variable
 from ..core import indexing
@@ -110,12 +109,16 @@ class H5NetCDFStore(WritableCFDataStore, DataStorePickleMixin):
         with self.ensure_open(autoclose=True):
             encoding = {}
             encoding['unlimited_dims'] = {
-                k for k, v in self.ds.dimensions.items() if v.isunlimited()}
+                k for k, v in self.ds.dimensions.items() if v is None}
         return encoding
 
-    def set_dimension(self, name, length):
+    def set_dimension(self, name, length, is_unlimited=False):
         with self.ensure_open(autoclose=False):
-            self.ds.createDimension(name, size=length)
+            if is_unlimited:
+                self.ds.createDimension(name, size=None)
+                self.ds.resize_dimension(name, length)
+            else:
+                self.ds.createDimension(name, size=length)
 
     def set_attribute(self, key, value):
         with self.ensure_open(autoclose=False):
@@ -144,7 +147,6 @@ class H5NetCDFStore(WritableCFDataStore, DataStorePickleMixin):
                     'chunksizes', 'fletcher32']:
             if key in encoding:
                 kwargs[key] = encoding[key]
-
         nc4_var = self.ds.createVariable(name, dtype, variable.dims,
                                          fill_value=fill_value, **kwargs)